fix: add matching types to resolve warnings

+++ main/main.cpp
@@ -11,7 +11,7 @@ int main(int argc, char **argv) {
 
   std::vector<RecommendedProduct> products;
   std::vector<std::string> clients = std::vector<std::string>(3, "");
-  int k;
+  long unsigned int k;
 
   std::cout << "Digite o codigo para o cliente 1: ";
   std::cin >> clients[0];
@@ -33,10 +33,10 @@ int main(int argc, char **argv) {
   SimilarityModule similarity(module.get_products_size(), module.get_client_purchase_history(), 1);
   RecommendationModule recommendation(similarity);
 
-  for(int i = 0; i < clients.size(); i++) {
+  for(unsigned int i = 0; i < clients.size(); i++) {
     std::cout << "Cliente " << i+1 << std::endl;
     products = recommendation.get_recommendations(module.get_client_id_by_client_code(clients[i]), k);
-    for(int j = 0; j < products.size(); j++) {
+    for(unsigned int j = 0; j < products.size(); j++) {
       std::cout << module.get_product_name_by_id(products[j].product_id) << " | Ranking Value: "<<  products[j].ranking_score << std::endl;
     }
     std::cout << std::endl;
